refactor: make an ai submodule
