+++ tests/index.html
@@ -1,4 +1,4 @@
-<!doctype html>
+<!DOCTYPE html>
 <html>
   <head>
     <title>Tests | Thirteen</title>
@@ -21,7 +21,8 @@
     <script type="module" src="./test_game.js"></script>
     <script type="module" src="./test_player.js"></script>
     <script type="module" src="./test_ui.js"></script>
-    <script type="module" src="./test_ai.js"></script>
+    <script type="module" src="./ai/test_base.js"></script>
+    <script type="module" src="./ai/test_lowest_card.js"></script>
     <script type="module" src="./test_app.js"></script>
     <script type="module" src="./test_analytics.js"></script>
     <script type="module">
