+++ tests/mocks.js
@@ -1,4 +1,4 @@
-import { AI } from "../src/app/ai.js";
+import { AI } from "../src/app/ai/index.js";
 import { Card } from "../src/app/deck.js";
 import { Game } from "../src/app/game.js";
 import { Player } from "../src/app/player.js";
